Add XdgDocumentsPortal
The Lookup, Info and List methods are not implemented - these calls are not available from inside a sandbox and are thus not likely to be used in Dart applications.

The AddNamed method is not implemented, this requires a file descriptor for a directory, which is not possible in Dart currently (files only).

Fixes https://github.com/canonical/xdg_desktop_portal.dart/issues/26